+++ kermit/k95/ckuus2.c
@@ -733,9 +733,13 @@ static char * hmxygui[] = {
 "  session configuration.  Once disabled the toolbar functions cannot be",
 "  re-enabled.",
 " ",
+"SET GUI STATUSBAR { ON, OFF }",
+"  Shows or hides the statusbar. Only works if K95G was not started with",
+"  the --nostatusbar or --nobars command line options."
+" ",
 "SET GUI TOOLBAR VISIBLE { ON, OFF }",
 "  Shows or hides the toolbar. Only works if K95G was not started with ",
-"  the --notoolbar command line option."
+"  the --notoolbar or --nobars command line options."
 " ",
 "SET GUI WINDOW POSITION x y",
 "  Moves the C-Kermit window to the given X,Y coordinates, pixels from top",