+++ kermit/k95/kui/ikui.h
@@ -36,6 +36,8 @@
 #define KUI_GUI_MENUBAR     1019
 #define KUI_GUI_CLOSE       1020
 #define KUI_GUI_TOOLBAR_VIS 1021
+#define KUI_GUI_STATBAR_VIS 1022
+
 
 struct _kui_init {
     char pos_init;		/* boolean if position set */